refactor(models): recombine into one module
keep it simple, especially with the tests now in a class-based structure
and able to encapsulate and reuse fixture data
