--- cdt_identity/models/__init__.py
@@ -1,7 +1,0 @@
-from .config import IdentityGatewayConfig
-from .request import ClaimsVerificationRequest
-
-__all__ = [
-    "IdentityGatewayConfig",
-    "ClaimsVerificationRequest",
-]
